+++ bc-gateway/bc-gateway-app/src/main/kotlin/co/nilin/opex/bcgateway/app/dto/TokenRequest.kt
@@ -1,13 +1,15 @@
 package co.nilin.opex.bcgateway.app.dto
 
+import java.math.BigDecimal
+
 data class TokenRequest(
     val symbol: String?,
     val chain: String?,
     val isToken: Boolean,
     val tokenName: String?,
     val tokenAddress: String?,
-    val withdrawFee: Double,
-    val minimumWithdraw: Double,
+    val withdrawFee: BigDecimal,
+    val minimumWithdraw: BigDecimal,
     val isWithdrawEnabled: Boolean = true,
     val decimal: Int = 18
 )
