+++ matching-gateway/matching-gateway-app/src/main/kotlin/co/nilin/opex/matching/gateway/app/inout/PairConfig.kt
@@ -1,9 +1,11 @@
 package co.nilin.opex.matching.gateway.app.inout
 
+import java.math.BigDecimal
+
 class PairConfig(
     val pair: String,
     val leftSideWalletSymbol: String, //can be same as pair left side
     val rightSideWalletSymbol: String, //can be same as pair right side
-    val rightSideFraction: Double,
-    val leftSideFraction: Double
+    val rightSideFraction: BigDecimal,
+    val leftSideFraction: BigDecimal
 )
